Clarify retry logic
Improve the separation between Fabric logic and the job queue implementation details
