+++ asset-transfer-basic/rest-api-typescript/src/config.ts
@@ -3,12 +3,12 @@
  *
  * The sample REST server can be configured using the environment variables
  * documented below
- * 
+ *
  * In a local development environment, these variables can be loaded from a
  * .env file by starting the server with the following command:
- * 
+ *
  *   npm start:dev
- * 
+ *
  * The scripts/generateEnv.sh script can be used to generate a suitable .env
  * file for the Fabric Test Network
  */
