+++ asset-transfer-basic/rest-api-typescript/src/fabric.spec.ts
@@ -11,7 +11,6 @@ import {
   submitTransaction,
   getBlockHeight,
   getTransactionValidationCode,
-  processSubmitTransactionJob,
 } from './fabric';
 import * as config from './config';
 
@@ -34,7 +33,6 @@ import * as fabricProtos from 'fabric-protos';
 
 import { MockProxy, mock } from 'jest-mock-extended';
 import Long from 'long';
-import { Job } from 'bullmq';
 
 jest.mock('./config');
 jest.mock('fabric-network', () => {
@@ -114,133 +112,6 @@ describe('Fabric', () => {
         assetContract: mockBasicContract,
         qsccContract: mockSystemContract,
       });
-    });
-  });
-
-  describe('processSubmitTransactionJob', () => {
-    const mockContracts = new Map<string, Contract>();
-    const mockPayload = Buffer.from('MOCK PAYLOAD');
-    const mockSavedState = Buffer.from('MOCK SAVED STATE');
-    let mockTransaction: MockProxy<Transaction>;
-    let mockContract: MockProxy<Contract>;
-    let mockJob: MockProxy<Job>;
-
-    beforeEach(() => {
-      mockTransaction = mock<Transaction>();
-      mockTransaction.getTransactionId.mockReturnValue('mockTransactionId');
-
-      mockContract = mock<Contract>();
-      mockContract.createTransaction
-        .calledWith('txn')
-        .mockReturnValue(mockTransaction);
-      mockContract.deserializeTransaction
-        .calledWith(mockSavedState)
-        .mockReturnValue(mockTransaction);
-      mockContracts.set('mockMspid', mockContract);
-
-      mockJob = mock<Job>();
-    });
-
-    it('gets job result with no error or payload if no contract is available for the required mspid', async () => {
-      mockJob.data = {
-        mspid: 'missingMspid',
-      };
-
-      const jobResult = await processSubmitTransactionJob(
-        mockContracts,
-        mockJob
-      );
-
-      expect(jobResult).toStrictEqual({
-        transactionError: undefined,
-        transactionPayload: undefined,
-      });
-    });
-
-    it('gets a job result containing a payload if the transaction was successful first time', async () => {
-      mockJob.data = {
-        mspid: 'mockMspid',
-        transactionName: 'txn',
-        transactionArgs: ['arg1', 'arg2'],
-      };
-      mockTransaction.submit
-        .calledWith('arg1', 'arg2')
-        .mockResolvedValue(mockPayload);
-
-      const jobResult = await processSubmitTransactionJob(
-        mockContracts,
-        mockJob
-      );
-
-      expect(jobResult).toStrictEqual({
-        transactionError: undefined,
-        transactionPayload: Buffer.from('MOCK PAYLOAD'),
-      });
-    });
-
-    it('gets a job result containing a payload if the transaction was successfully rerun using saved transaction state', async () => {
-      mockJob.data = {
-        mspid: 'mockMspid',
-        transactionName: 'txn',
-        transactionArgs: ['arg1', 'arg2'],
-        transactionState: mockSavedState,
-      };
-      mockTransaction.submit
-        .calledWith('arg1', 'arg2')
-        .mockResolvedValue(mockPayload);
-
-      const jobResult = await processSubmitTransactionJob(
-        mockContracts,
-        mockJob
-      );
-
-      expect(jobResult).toStrictEqual({
-        transactionError: undefined,
-        transactionPayload: Buffer.from('MOCK PAYLOAD'),
-      });
-    });
-
-    it('gets a job result containing an error message if the transaction fails but cannot be retried', async () => {
-      mockJob.data = {
-        mspid: 'mockMspid',
-        transactionName: 'txn',
-        transactionArgs: ['arg1', 'arg2'],
-        transactionState: mockSavedState,
-      };
-      mockTransaction.submit
-        .calledWith('arg1', 'arg2')
-        .mockRejectedValue(
-          new Error(
-            'Failed to get transaction with id txn, error Entry not found in index'
-          )
-        );
-
-      const jobResult = await processSubmitTransactionJob(
-        mockContracts,
-        mockJob
-      );
-
-      expect(jobResult).toStrictEqual({
-        transactionError:
-          'TransactionNotFoundError: Failed to get transaction with id txn, error Entry not found in index',
-        transactionPayload: undefined,
-      });
-    });
-
-    it('throws an error if the transaction fails but can be retried', async () => {
-      mockJob.data = {
-        mspid: 'mockMspid',
-        transactionName: 'txn',
-        transactionArgs: ['arg1', 'arg2'],
-        transactionState: mockSavedState,
-      };
-      mockTransaction.submit
-        .calledWith('arg1', 'arg2')
-        .mockRejectedValue(new Error('MOCK ERROR'));
-
-      await expect(async () => {
-        await processSubmitTransactionJob(mockContracts, mockJob);
-      }).rejects.toThrow('MOCK ERROR');
     });
   });
